ml-training/requirements.txt
tensorflow>=2.17.0
pillow>=10.0.0
numpy>=1.24.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
